+++ CHANGELOG.md
@@ -17,6 +17,8 @@ For older change log history see the [historic changelog](HISTORIC_CHANGELOG.md)
 - Added CMForestDiscovery resource
 - Added ConvertTo-CimCMScheduleString to the ResourceHelper
 - Added CMClientStatusSettings Resource
+- Added CMPullDistributionPoint Resource
+- Added ConvertTo-AnyCimInstance to the ResourceHelper
 
 ### Changed